feat(loader): start work on loading files into package graph
This commit will start the implementation of the process of loading
the main files and dependencies into the package graph. It requires
parsing and processing all messages in each file, generating use
references and import references.

In this commit, the process only handles loading the main files,
without handling any sort of import. Currently, the mapping from
names to IDs are stored in the loader state, another implementation
possibility is to store these maps in the graph it self, which would
enable us to have get_by_name functions.